StiCky NoTeS
This is a sticky note project built on HTML5 ,CSS3 and JavaScript.
Through this you can make your notes delete it.
this project is a PWA i.e Progressive Web App .
It can be installed on your device and works just like the any other normal app on your device even in Offline Mode.